Restructured the project to follow feature first
