+++ api/features/user/controller.go
@@ -1,0 +1,15 @@
+package user
+
+import (
+	"context"
+	db "go-todo/db/sqlc"
+)
+
+type UserController struct {
+	db *db.Queries
+	ctx context.Context
+}
+
+func NewController(db *db.Queries, ctx context.Context) *UserController {
+	return &UserController{db:db, ctx: ctx}
+}
