+++ api/util/mycontext/context.go
@@ -1,7 +1,10 @@
-package util
+package mycontext
 
 import (
 	"errors"
+	"fmt"
+	"go-todo/gterrors"
+	"go-todo/util/txtutil"
 
 	"github.com/gin-gonic/gin"
 )
@@ -29,4 +32,28 @@ func GetTokenVariables(ctx *gin.Context) (userID string, username string, isAdmi
 	}
 
 	return userID, username, isAdmin, nil
+}
+
+func CtxAddGtInternalError(message, file string, line int, err error, c *gin.Context) {
+	errToAdd := err
+	if message != "" {
+		errToAdd = fmt.Errorf("%v: %w", message, err)
+	}
+	c.Error(
+		gterrors.NewGtInternalError(
+			errToAdd,
+			txtutil.AddLineNumberToFileName(file,line),
+			500,
+		),
+	).SetType(gterrors.GetGinErrorType())
+}
+
+// Check the body format against the schema. 'payload' should be like
+// &schemas.<MyBodySchema>
+func ShouldBindBodyWithJSON(payload any, c *gin.Context) bool {
+	if err := c.ShouldBindBodyWithJSON(&payload); err != nil {
+		c.Error(err).SetType(gin.ErrorTypeBind)
+		return false
+	}
+	return true
 }
