--- api/util/string_formatters.go
@@ -1,7 +1,0 @@
-package util
-
-import "fmt"
-
-func GetFileNameWithLine(fileName string, lineNumber int) string {
-	return fmt.Sprintf("%v: %d", fileName, lineNumber)
-}
