+++ api/util/txtutil/txtutil.go
@@ -1,0 +1,7 @@
+package txtutil
+
+import "fmt"
+
+func AddLineNumberToFileName(fileName string, lineNumber int) string {
+	return fmt.Sprintf("%v: %d", fileName, lineNumber)
+}
